+++ LBHTenancyAPI/UseCases/V1/TenancyDetailsForRef.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using LBHTenancyAPI.Gateways;
+using LBHTenancyAPI.Gateways.V1;
 using LBHTenancyAPI.UseCases.V1;
 
 namespace LBHTenancyAPI.UseCases
